Refractored code for data driven tests and PEML Test into two different files
